QA-chatbot-mcp-servers/requirements.txt
streamlit
google-api-python-client
python-dotenv 